+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add cilium network policies to allow kube-api access.
+- Add network policies for egress.
+
 ## [1.7.0] - 2022-10-06
 
 ### Added